ISSUE #3878 - rename settingsForm to settingsModal
