+++ frontend/src/v5/ui/routes/dashboard/projects/federations/createFederationForm/createFederationSettings/createFederationSettings.component.tsx
@@ -21,7 +21,7 @@ import { FormSelect, FormTextField } from '@controls/inputs/formInputs.component
 import { MenuItem } from '@mui/material';
 import { useFormContext } from 'react-hook-form';
 import { FormattedMessage } from 'react-intl';
-import { SectionTitle } from '../../../settingsForm/settingsForm.styles';
+import { SectionTitle } from '../../../settingsModal/settingsModal.styles';
 import { HalfWidth } from './createFederationSettings.styles';
 
 export const CreateFederationFormSettings = () => {
